fix: Only create env file if not already exists
This keeps it from accidentally overwriting an existing env file if the user is just updating their setup or set up their env file manually ahead of time.

+++ bin/setup.sh
@@ -48,7 +48,12 @@ pip install -Ur requirements.txt
 deactivate
 
 # Create .env file
-cp .template.env .env
+if [[ -f .env ]]; then
+  echo ".env file already created, skipping"
+else
+  echo "Creating .env file from template"
+  cp .template.env .env
+fi
 
 # Creates log folder
 echo "Creating log folder"
